Add hdl and hls PID to lib
